==> Codes/3 KNN for MVP prediction.ipynb <==
#Setup

##Google Drive

!pip install sportsreference
from pydrive.auth import GoogleAuth
from pydrive.drive import GoogleDrive
from google.colab import auth
from oauth2client.client import GoogleCredentials
import os

# 1. Authenticate and create the PyDrive client.
auth.authenticate_user()
gauth = GoogleAuth()
gauth.credentials = GoogleCredentials.get_application_default()
drive = GoogleDrive(gauth)  

##Import Packages

import pickle
import pandas as pd
import numpy as np
from sportsreference.nba.teams import Teams
from sportsreference.nba.roster import Roster
from sportsreference.nba.roster import Player
from sportsreference.nba.boxscore import Boxscore
from sklearn.utils import shuffle
from sklearn.preprocessing import StandardScaler
from sklearn.neighbors import NearestNeighbors
from statsmodels.formula.api import logit
from sklearn.metrics import roc_curve
from sklearn.metrics import auc
import seaborn as sns
from matplotlib import pyplot as plt
import warnings
warnings.filterwarnings("ignore")
np.random.seed(123)

##Download files

# choose a local (colab) directory to store the data.
local_download_path = os.path.expanduser('~/data/pickle/')
try:
  os.makedirs(local_download_path)
except: pass

# 2. Auto-iterate using the query syntax
#    https://developers.google.com/drive/v2/web/search-parameters
file_list = drive.ListFile(
    {'q': "'1e8GS0L0xUXQDgiRorx__FQViQjHNza7c' in parents"}).GetList()

for f in file_list:
  # 3. Create & download by id.
  print('title: %s, id: %s' % (f['title'], f['id']))
  fname = os.path.join(local_download_path, f['title'])
  print('downloading to {}'.format(fname))
  f_ = drive.CreateFile({'id': f['id']})
  f_.GetContentFile(fname)

#Load Data

##Load data from pickle

# retrieve players' stats data from pickle files
all_players_data_all_season_file = '/root/data/pickle/all_players_data_all_season_processed.pkl'
with open(all_players_data_all_season_file, 'rb') as f:
  all_players_data = pickle.load(f)

#Data preparation

##Season data

seasons_consider = ['2013-14', '2014-15', '2015-16', '2016-17', '2017-18', '2018-19', '2019-20']
all_players_data = all_players_data.loc[all_players_data['season'].isin(seasons_consider), :]

##Set index

all_players_data = all_players_data.set_index(['season', 'player_name', 'player_id'])

##Select columns

# Performance columns
perf_cols = [
            # 'assist_percentage',
            'assists',
            # 'block_percentage',
            # 'blocks',
            'box_plus_minus',
            'center_percentage',
            # 'defensive_box_plus_minus',
            # 'defensive_rebound_percentage',
            # 'defensive_rebounds',
            # 'defensive_win_shares',
            # 'dunks',
            'effective_field_goal_percentage',
            # 'field_goal_attempts',
            'field_goal_perc_sixteen_foot_plus_two_pointers',
            'field_goal_perc_ten_to_sixteen_feet',
            'field_goal_perc_three_to_ten_feet',
            # 'field_goal_perc_zero_to_three_feet',
            'field_goal_percentage',
            'field_goals',
            # 'free_throw_attempt_rate',
            'free_throw_attempts',
            'free_throw_percentage',
            'free_throws',
            'games_played',
            'games_started',
            # 'half_court_heaves',
            # 'half_court_heaves_made',
            'minutes_played',
            # 'offensive_box_plus_minus',
            # 'offensive_rebound_percentage',
            # 'offensive_rebounds',
            # 'offensive_win_shares',
            # 'percentage_field_goals_as_dunks',
            'percentage_of_three_pointers_from_corner',
            # 'percentage_shots_three_pointers',
            # 'percentage_shots_two_pointers',
            # 'percentage_sixteen_foot_plus_two_pointers',
            # 'percentage_ten_to_sixteen_footers',
            'percentage_three_to_ten_footers',
            # 'percentage_zero_to_three_footers',
            # 'personal_fouls',
            'player_efficiency_rating',
            'point_guard_percentage',
            'points',
            'power_forward_percentage',
            # 'shooting_distance',
            'shooting_guard_percentage',
            'small_forward_percentage',
            # 'steal_percentage',
            # 'steals',
            # 'three_point_attempt_rate',
            # 'three_point_attempts',
            # 'three_point_percentage',
            # 'three_point_shot_percentage_from_corner',
            # 'three_pointers',
            # 'three_pointers_assisted_percentage',
            # 'total_rebound_percentage',
            'total_rebounds',
            'true_shooting_percentage',
            'turnover_percentage',
            'turnovers',
            'two_point_attempts',
            # 'two_point_percentage',
            'two_pointers',
            'two_pointers_assisted_percentage',
            'usage_percentage',
            'win_shares',
            'win_shares_per_48_minutes'
              ]
all_players_data = all_players_data[perf_cols]

##Train, cross-validation, and test split

# Train
Train = all_players_data.loc[['2013-14', '2014-15', '2015-16', '2016-17'],:]
# Val
Val1 = all_players_data.loc[['2017-18'],:]
Val2 = all_players_data.loc[['2018-19'],:]
# Test
Test = all_players_data.loc[['2019-20'],:]

#KNN for MVP determination

##Scaling variables

scaler = StandardScaler()

# Train
scaler.fit(Train)

# Training data
Train_std = scaler.transform(Train)

# Validation data
Val1_std = scaler.transform(Val1)
Val2_std = scaler.transform(Val2)

# Test data
Test_std = scaler.transform(Test)

##Define Model

model = NearestNeighbors(n_neighbors = 10)

##Calculating cetroid of previous MVPs

index1 = Train.index.get_loc(('2013-14', 'Kevin Durant', 'duranke01'))
index2 = Train.index.get_loc(('2014-15', 'Stephen Curry', 'curryst01'))
index3 = Train.index.get_loc(('2015-16', 'Stephen Curry', 'curryst01'))
index4 = Train.index.get_loc(('2016-17', 'Russell Westbrook', 'westbru01'))
centroid = Train_std[[index1,index2,index3,index4],:].mean(axis = 0)

##MVP prediction for season 2017-18 (Validation 1)

model.fit(Val1_std) 
distance, indices = model.kneighbors([centroid])

print(Val1.iloc[indices[0],:].index)

##MVP prediction for season 2018-19 (Validation 2)



model.fit(Val2_std) 
distance, indices = model.kneighbors([centroid])

print(Val2.iloc[indices[0],:].index)

##MVP prediction for current season (Testing)

model.fit(Test_std) 
distance, indices = model.kneighbors([centroid])

print(Test.iloc[indices[0],:].index)

# distance_ind = (centroid - Val2_std[53,:])**2
# distance_ind_perc = distance_ind/distance_ind.sum()*100
# d_ = dict()
# for i, j in zip(Val2.columns, distance_ind_perc):
#   d_[i] = j
# pd.set_option("display.max_rows",None)
# print(pd.DataFrame(d_, index=[1]).T)
# pd.set_option("display.max_rows",60)
